% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\docType{data}
\name{cr_grid_1km}
\alias{cr_grid_1km}
\alias{cr_grid_2km}
\alias{cr_grid_4km}
\alias{cr_grid_8km}
\title{Spatial grids for Costa Rica of 1km, 2km, 4km and 8km}
\format{
An sf object (EPSG:5367)
}
\usage{
cr_grid_1km
cr_grid_2km
cr_grid_4km
cr_grid_8km
}
\description{
Spatial grids for Costa Rica of 1km, 2km, 4km and 8km.
}
\keyword{datasets}
