% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crgrids-package.R
\docType{package}
\name{crgrids-package}
\alias{crgrids}
\alias{crgrids-package}
\title{crgrids: crgrids: Spatial grids for Costa Rica}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Contains spatial grids for Costa Rica (1km, 2km, 4km, 8km).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ManuelSpinola/crgrids}
  \item \url{https://manuelspinola.github.io/crgrids/}
}

}
\author{
\strong{Maintainer}: Manuel Spínola \email{mspinola10@gmail.com} (\href{https://orcid.org/0000-0002-7839-1908}{ORCID})

Authors:
\itemize{
  \item Guido Saborío (\href{https://orcid.org/0000-0001-9490-2459}{ORCID})
}

}
\keyword{internal}
